Fatiri

See merge request enigma-camp/class-b/final-project/kelo/mang-tix-be!34

+++ src/main/java/com/enigma/services/impl/BookingServiceImpl.java
@@ -1,7 +1,7 @@
 package com.enigma.services.impl;
 
+import com.enigma.constanta.BookingConstant;
 import com.enigma.constanta.MessageConstant;
-import com.enigma.constanta.StringConstant;
 import com.enigma.entity.*;
 import com.enigma.exception.ForbiddenException;
 import com.enigma.exception.NotFoundException;
@@ -11,10 +11,13 @@ import com.enigma.services.EventService;
 import com.enigma.services.TicketService;
 import com.enigma.services.UserService;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.data.domain.Page;
+import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
 
 import java.math.BigDecimal;
-import java.util.*;
+import java.util.ArrayList;
+import java.util.List;
 
 @Service
 public class BookingServiceImpl implements BookingService {
@@ -28,29 +31,39 @@ public class BookingServiceImpl implements BookingService {
     @Autowired
     EventService eventService;
 
+
     @Override
     public Booking booking(Booking booking) {
         User user = userService.getUserById(booking.getUserIdTransient());
         booking.setUser(user);
-        booking.setPaymentStatus(false);
-        BigDecimal sumSubtotal;
+        List<TicketCode> listTicketAvailable = new ArrayList<>();
+        Integer sumTicketAvailable =0;
 
-        for (BookingDetail bookingDetail : booking.getBookingDetailList()) {
-            bookingDetail.setBooking(booking);
-            Ticket ticket = ticketService.getTicketById(bookingDetail.getTicketIdTransient());
-            bookingDetail.setTicket(ticket);
-
-            ticketService.deduct(bookingDetail.getTicket().getId(), bookingDetail.getQuantity());
-            sumSubtotal = ticket.getPrice().multiply(new BigDecimal(bookingDetail.getQuantity()));
-            bookingDetail.setSubtotal(sumSubtotal);
+        for (BookingDetail bookingDetailList: booking.getBookingDetailList()) {
+            Ticket ticket = ticketService.getTicketById(bookingDetailList.getTicketIdTransient());
+            bookingDetailList.setTicket(ticket);
+            for (TicketCode ticketCodeListOne: bookingDetailList.getTicket().getTicketCodes()) {
+                if (ticketCodeListOne.getAvailable()){
+                    listTicketAvailable.add(ticketCodeListOne);
+                    for (int i = 1; i <=listTicketAvailable.size() ; i++) {
+                        sumTicketAvailable = i;
+                        if (bookingDetailList.getQuantity()<=sumTicketAvailable){
+                            ticketCodeListOne.setAvailable(true);
+                            bookingDetailList.setSubtotal(bookingDetailList.getTicket().getPrice().multiply(new BigDecimal(bookingDetailList.getQuantity())));
+                        }
+                    }
+                }
+                throw new ForbiddenException(BookingConstant.TICKET_NOT_AVAILABLE);
+            }
 
         }
+        System.out.println(sumTicketAvailable);
         return bookingRepository.save(booking);
     }
 
     @Override
-    public List<Booking> getAllBookingData() {
-        return bookingRepository.findAll();
+    public Page<Booking> getAllBookingData(Pageable pageable) {
+        return bookingRepository.findAll(pageable);
     }
 
     @Override
@@ -63,11 +76,8 @@ public class BookingServiceImpl implements BookingService {
 
     @Override
     public void deleteBookingDataById(String bookingId) {
-        Booking booking = getBookingById(bookingId);
-        for (BookingDetail bookingDetail : booking.getBookingDetailList()) {
-            String idTicket = bookingDetail.getTicket().getId();
-            Ticket ticket = ticketService.getTicketById(idTicket);
-            ticket.setQuantity(ticket.getQuantity() + bookingDetail.getQuantity());
+        if (!bookingRepository.findById(bookingId).isPresent()) {
+            throw new NotFoundException(String.format(MessageConstant.ID_BOOKING_NOT_FOUND, bookingId));
         }
         bookingRepository.deleteById(bookingId);
     }
